Refactor: Move ElementTypes12 into its own file, and pass it from the top-level xliffToJs12()
