+++ inline-elements/typeToTagMaps.js
@@ -1,0 +1,9 @@
+function elementTypeToTag(type, types) {
+  return types.elementTypeToTagMap[type];
+}
+exports.elementTypeToTag = elementTypeToTag;
+
+function tagToElementType(tagName, types) {
+  return types.tagToElementTypeMap[tagName];
+}
+exports.tagToElementType = tagToElementType;
