+++ test/xml-js/objectToXml.test.js
@@ -1,7 +1,7 @@
 const expect = require('expect.js');
-const ElementTypes = require('../../inline-elements/ElementTypes').ElementTypes;
-const ElementTypes12 = require('../../inline-elements/ElementTypes').ElementTypes12;
-const elementTypeToTag = require('../../inline-elements/ElementTypes').elementTypeToTag;
+const ElementTypes = require('../../inline-elements/ElementTypes');
+const ElementTypes12 = require('../../inline-elements/ElementTypes12');
+const elementTypeToTag = require('../../inline-elements/typeToTagMaps').elementTypeToTag;
 const makeElement = require('../../xml-js/objectToXml').makeElement;
 const makeText = require('../../xml-js/objectToXml').makeText;
 const makeValue = require('../../xml-js/objectToXml').makeValue;
