+++ xliff12ToJs.js
@@ -1,4 +1,5 @@
 const convert = require('xml-js');
+const ElementTypes12 = require('./inline-elements/ElementTypes12');
 const extractValue = require('./xml-js/xmlToObject').extractValue;
 
 function xliff12ToJs(str, cb) {
@@ -39,7 +40,7 @@ function xliff12ToJs(str, cb) {
           case 'source':
           case 'target':
           case 'note':
-            unit[element.name] = extractValue(element.elements);
+            unit[element.name] = extractValue(element.elements, ElementTypes12);
             break;
         }
